@@ -1,6 +1,6 @@
 {
-  "name": "vida",
-  "version": "0.0.1",
+  "name": "@alanmorel/vida",
+  "version": "0.0.6",
   "description": "",
   "main": "dist/vida.ssr.js",
   "browser": "dist/vida.esm.js",
@@ -19,24 +19,26 @@
   },
   "dependencies": {},
   "devDependencies": {
-    "@babel/core": "^7.9.0",
-    "@babel/preset-env": "^7.9.5",
-    "@rollup/plugin-alias": "^2.2.0",
-    "@rollup/plugin-commonjs": "^11.1.0",
-    "@rollup/plugin-replace": "^2.3.2",
-    "@vue/cli-plugin-babel": "^4.3.1",
-    "@vue/cli-service": "^4.3.1",
+    "@babel/core": "^7.11.4",
+    "@babel/preset-env": "^7.11.0",
+    "@rollup/plugin-alias": "^3.1.1",
+    "@rollup/plugin-babel": "^5.2.0",
+    "@rollup/plugin-commonjs": "^14.0.0",
+    "@rollup/plugin-node-resolve": "^9.0.0",
+    "@rollup/plugin-replace": "^2.3.3",
+    "@vue/cli-plugin-babel": "^4.5.4",
+    "@vue/cli-service": "^4.5.4",
+    "@vue/compiler-sfc": "^3.0.0-rc.9",
     "cross-env": "^7.0.2",
     "minimist": "^1.2.5",
-    "rollup": "^2.7.3",
-    "rollup-plugin-babel": "^4.4.0",
-    "rollup-plugin-terser": "^5.3.0",
-    "rollup-plugin-vue": "^5.1.6",
-    "vue": "^2.6.11",
-    "vue-template-compiler": "^2.6.11"
+    "rollup": "^2.26.5",
+    "rollup-plugin-postcss": "^3.1.6",
+    "rollup-plugin-terser": "^7.0.0",
+    "rollup-plugin-vue": "^6.0.0-beta.10",
+    "vue": "^3.0.0-rc.9"
   },
   "peerDependencies": {
-    "vue": "^2.6.11"
+    "vue": "^3.0.0-rc.9"
   },
   "engines": {
     "node": ">=10"
