@@ -1,39 +1,12 @@
-// Import vue components
-import * as components from "@/components";
+// iife/cjs usage extends esm default export - so import it all
+// import * as components from '@/lib-components/index';
+import plugin, * as components from '@/entry.esm';
 
-// install function executed by Vue.use()
-const install = function installVida(Vue) {
-    if (install.installed) { 
-        return;
-    }
-    install.installed = true;
-    Object.entries(components).forEach(([componentName, component]) => {
-        Vue.component(componentName, component);
-    });
-};
+// Attach named exports directly to component. IIFE/CJS will
+// only expose one global var, with named exports exposed as properties of
+// that global var (eg. VivintIcon.iconList)
+Object.entries(components).forEach(([componentName, component]) => {
+  if (componentName !== 'default') plugin[componentName] = component;
+});
 
-// Create module definition for Vue.use()
-const plugin = {
-    install
-};
-
-// To auto-install on non-es builds, when vue is found
-// eslint-disable-next-line no-redeclare
-/* global window, global */
-if ("false" === process.env.ES_BUILD) {
-    let GlobalVue = null;
-    if (typeof window !== "undefined") {
-        GlobalVue = window.Vue;
-    } else if (typeof global !== "undefined") {
-        GlobalVue = global.Vue;
-    }
-    if (GlobalVue) {
-        GlobalVue.use(plugin);
-    }
-}
-// Default export is library as a whole, registered via Vue.use()
 export default plugin;
-
-// To allow individual component use, export components
-// each can be registered via Vue.component()
-export * from "@/components";
