@@ -1,6 +1,8 @@
 <script>
-export default {
-  name: "VidaSample", // vue component name
+import { defineComponent } from 'vue';
+
+export default defineComponent({
+  name: 'VidaSample', // vue component name
   data() {
     return {
       counter: 5,
@@ -14,30 +16,30 @@ export default {
   computed: {
     changedBy() {
       const { message } = this;
-      if (!message.action) return "initialized";
-      return `${message?.action} ${message.amount ?? ""}`.trim();
+      if (!message.action) return 'initialized';
+      return `${message?.action} ${message.amount ?? ''}`.trim();
     },
   },
   methods: {
     increment(arg) {
-      const amount = (typeof arg !== "number") ? 1 : arg;
+      const amount = (typeof arg !== 'number') ? 1 : arg;
       this.counter += amount;
-      this.message.action = "incremented by";
+      this.message.action = 'incremented by';
       this.message.amount = amount;
     },
     decrement(arg) {
-      const amount = (typeof arg !== "number") ? 1 : arg;
+      const amount = (typeof arg !== 'number') ? 1 : arg;
       this.counter -= amount;
-      this.message.action = "decremented by";
+      this.message.action = 'decremented by';
       this.message.amount = amount;
     },
     reset() {
       this.counter = this.initCounter;
-      this.message.action = "reset";
+      this.message.action = 'reset';
       this.message.amount = null;
     },
   },
-};
+});
 </script>
 
 <template>
